pyproject.toml
[project]
name = "complianceqapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-monitor-opentelemetry>=1.8.6",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.129.0",
    "firecrawl-py>=4.16.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "opentelemetry-instrumentation-httpx>=0.60b0",
    "opentelemetry-instrumentation-requests>=0.60b0",
    "opentelemetry-instrumentation-urllib3>=0.60b0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
    "yt-dlp>=2026.2.4",
    "opencv-python>=4.10.0",
    "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
addopts = "-q"
